+++ core-ng/src/main/java/core/framework/db/Database.java
@@ -1,8 +1,5 @@
 package core.framework.db;
 
-import core.framework.internal.log.ActionLog;
-import core.framework.internal.log.LogManager;
-
 import java.util.List;
 import java.util.Optional;
 
@@ -10,26 +7,6 @@ import java.util.Optional;
  * @author neo
  */
 public interface Database {
-    /**
-     * for expected slow sql, such as query with small table or db chooses different query plan,
-     * disable warning then enable after<p>
-     * e.g.
-     * <blockquote><pre>
-     * Database.suppressSlowSQLWarning(false);
-     * database.select(...);
-     * Database.suppressSlowSQLWarning(true);
-     * </pre></blockquote>
-     *
-     * @param suppress whether suppress slow sql warning
-     */
-    static void suppressSlowSQLWarning(boolean suppress) {
-        // pass flag thru thread local to MySQLQueryInterceptor, and put in action log to reset for every action
-        ActionLog actionLog = LogManager.currentActionLog();
-        if (actionLog != null) {
-            actionLog.warningContext.suppressSlowSQLWarning = suppress;
-        }
-    }
-
     <T> List<T> select(String sql, Class<T> viewClass, Object... params);
 
     <T> Optional<T> selectOne(String sql, Class<T> viewClass, Object... params);