+++ core-ng/src/main/java/core/framework/db/QueryDiagnostic.java
@@ -1,9 +1,0 @@
-package core.framework.db;
-
-public interface QueryDiagnostic {
-    String sql();
-
-    boolean noGoodIndexUsed();
-
-    boolean noIndexUsed();
-}