+++ core-ng/src/main/java/core/framework/internal/db/inspector/QueryAnalyzer.java
@@ -1,0 +1,5 @@
+package core.framework.internal.db.inspector;
+
+public interface QueryAnalyzer {
+    QueryPlan explain(String sql, Object[] params);
+}