+++ core-ng/src/test/java/core/framework/db/QueryTest.java
@@ -2,9 +2,10 @@ package core.framework.db;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
 import org.mockito.Mockito;
-import org.mockito.MockitoAnnotations;
+import org.mockito.junit.jupiter.MockitoExtension;
 
 import java.util.List;
 
@@ -16,13 +17,13 @@ import static org.mockito.Mockito.verify;
 /**
  * @author neo
  */
+@ExtendWith(MockitoExtension.class)
 class QueryTest<T> {
     @Mock
     Query<T> query;
 
     @BeforeEach
     void createQuery() {
-        MockitoAnnotations.openMocks(this);
         Mockito.doCallRealMethod().when(query).in(any(), anyList());
     }